Lighbox is so much better now!
It's getting there, thanks to Jude. Now I just have to figure out why
there's such a huge right-hand margin on the images taking away space
from the captions. It does seem to be coming from the image container,
but it's not a margin or padding. Can't track it down.